Reduce CDS console output
These are pretty minor, but we output an error on startup if there is no
specified id and no contest.yaml (and hence we have to use the path to
default the id instead). This isn't really an error, so check if the file
exists before trying, and reduce it to warning.

I've also reduced:
  Finished scanning astana for changed resources in 94ms
to:
  Scanned astana for changes in 94ms
just to reduce the constant output a little bit.