Streaming, and resolved file ref should (only) be file
Found this while working on #1133.

When resolving a file reference, the only possible type should be file,
not the data (contents) of the file. The original choice of Object was
to allow submission reactions to be streamed before the reaction video
was done recording, but this path was changed years ago and the current
streaming code (sending reactions while they were still being recorded)
wasn't even working (and didn't support multiple streams).

This fixes the reaction streaming in ContestRESTService (moving it to a
common helper on HttpHelper) and then correctly types all resolved file
to File to improve typing/readability.