Trace when server provides image with duplicate metadata
The CDS will repeatedly download 'duplicate' images if the Contest API server
says the file is different than what the CDS determines it to be after
downloading. For example, if the server says it has an image that's 50x50,
the CDS downloads it, and reads it as 51x51, then the next time you restart
we'll have a 51x51 image, see that the source has a 'different' one, and
download it a second time. This will happen every time the CDS restarts,
leading to many duplicate images.

The most obvious solution is to fix the bug in determining image size or
file reference data in the CDS or the upstream server, but that takes time
and is hard to track down, so this just starts by logging whenever this
occurs to make it easier to track. In the future this code would make it
easy to take the next step and delete the file after download, but I want
to run it in a few contests before making that step. This change:

- Moves the logic of whether a file reference matches an existing one from
  PlaybackContest to FileReferenceList.
- Exposes the local file cache information from DiskContestSource.
- After downloading in RestContestSource, check if the file matches one we
  already have.

Most of #1125.