Remove extraneous photo property
While testing the Baku data I noticed that all teams and persons had a photo
property '"photo": []' even when they didn't have any photo. This shouldn't hurt
anything, but it's an extraneous/unnecessary property (minor memory/http overhead)
so I was curious.

The support for default (backup) photos shouldn't be static (it's per-contest) and
was always setting the photo property even when there was no default photo (99% of
the time). This change just adds some extra if statements to avoid setting the
property and optimize for the typical case where it's not set.