Stop background scanning after removing contest
While testing I noticed that when I removed a contest from cdsConfig.xml it
correctly disappeared from the web/API, but the background task was still
periodically scanning for resource changes in the background.

This just keeps track of the background task and correctly cleans it up
when the contest is closed.

I also improved the message so that when you remove a contest from the config
it says "Removing baku" instead of "Removing 87384679265".