Experimental support for simplified judgement types
Adds a new optional property to judgement types, 'simplified_judgement_type_id'.
If this is set, any judgements with this type are changed to the simplified
judgement type for other teams. (i.e. you can see the detail on your judgements,
but not for other teams)

To try this out, set up a contest with a bunch of submissions and judgements,
and at least one team account. Create a judgement-types.json in the contest
root by capturing the current output of /judgement-types (or just create the
content manually). Then add the new property to some of the judgements, e.g.:
  {"id":"OLE","name":"output limit","penalty":true,"solved":false,"simplified_judgement_type_id":"FAIL"},
  {"id":"PE","name":"presentation error","penalty":true,"solved":false,"simplified_judgement_type_id":"FAIL"},
  {"id":"WA","name":"wrong answer","penalty":true,"solved":false,"simplified_judgement_type_id":"FAIL"},

The id can point to one of the existing judgement types, or you can create a
new one for this purpose, e.g.:
  {"id":"FAIL","name":"just wrong","penalty":true,"solved":false}]

I didn't do any optimization yet, this is just an PoC to see if this is where
the Contest API should go.