Support for multiple teams and groups in clarification destinations
Adds support for the draft Contest API ability to send clarifications to multiple
teams and/or groups.

This updates the contest model and all client-side support to support multiple
teams and groups. The CDS, presentations, and all other tools will work when
pointing to a 2023-06 _or_ draft spec server.

By default the CDS will continue to serve the 2023-06 contest API though; if you
want the new event feed you need to set the ICPC_CONTEST_API=draft system property
as with all other draft spec changes.

CDS web pages will need to be updated as well, that will be a separate PR.