Replace deprecated Locale constructor
Since Java 19, all constructors of the Locale class have been deprecated.
Instead, you should use the Locale.of static constructors to create a new
Locale instance or use the Locale.Builder class to create a new Locale instance.

The easiest translation would be Locale.of but it's not available yet in the
targeted Java version used in CI.