Specify which registry to trust
The Docker binary defaults to docker.io as registry. Podman as
alternative doesn't and might ship with no default registry so can't
build this container.

By explicit specifying the registry the risk for inconsistent builds is also decreased
in case the GHA ever encounters a similar named container on the GHCR and would not use docker.