Only display the logo when there is space
Without it the whole line gets presses to make sure the whole
team/university name fits. When this is wider than the box we would
press the whole line including the displayed logo.

We now cheat and just not draw that logo giving the name a bit more
space and prevent the logo from being displayed out of ratio.