Default presentation for presentation client
Adds an option to set a default presentation for connected presentation clients. This
presentation is set before trying to connect to the CDS. If the CDS doesn't set any
presentation then this one is shown; otherwise whatever the CDS says takes over.

This avoids the standard ICPC Tools loading page, and allows you to have a default
w/o waiting for clients to connect and then setting it in the pres admin.

Fixes #955.