Fix lightmode for ContestFloorPresentation
The TextHelper would always set it's own color and not the one from the
Graphics.

Better would be if we can autowire the lightmode in there so we can make the
selection from here and optionally even get the current color from the graphics
if that's non standard so we pick the right one always.