Always prefer SVGs
There's a very small chance that an SVG just contains a low res image, but in all
other cases SVGs are better: the file download time is less, loading is less, and
generating an image of any exact size should be comparable to png scaling. This
change always picks an SVG as 'best' if one exists in a file reference list,
regardless of requested image size.