Scoreboard cell shading for scoring contests
Instead of strict green or red for success or failed submissions, this updates the
scoreboard cells for scoring contests with a problem max_score to shade the cells
anywhere between green (100% of max score) to red (0% of max score).

Recent changes normally flash, but I had to remove this for now as we'd have to go
transparent or pick a new colour, which is a bigger change.