Fix balloon printer and update SWT launcher
Adding the sentry dependency broke a few of the tool scripts a couple months ago.
Instead of the direct fix of adding sentry to all the scripts, I tried to simplify
and allow all of them to run against all jars in the lib folder (June 20 / build 1215),
but I had forgotten how the SWT launcher works.

This may be going further down the rabbit hole, but it simplifies how the SWT launcher
works as well, removes unused (and potentially confusing) manifest files, and fixes the
scripts to match.

# Conflicts:
#	PresAdmin/scripts/presAdmin.bat
#	PresAdmin/scripts/presAdmin.sh