Always keep team detail logo onscreen
If the team detail (intro) presentation has both photo and logo, the logo goes at
the top right of the photo. If the screen is too narrow (wider than it is high) the
logo can be drawn partially off the top of the screen. This just moves it down in
this case.