QR code generator util
Added a main() method to the QR code generator to make it easy to generate codes.
Just run it with two arguments: the QR code to generate and the filename to use.
An image with that QR code will be generated with that name in the working folder.

Also documented a few handy codes for our current scanners.