Remove disk scanning logging
When I added periodic disk scanning for contests I added tracing so that I could confirm
it was happening and see how long it was taking each time, especially on large contests.
I've never seen more than 10-20ms in dev or prod and the constant tracing even when
'nothing is happening' is annoying and adds up if you leave the contest running, so I'm
removing it.