Remove getFiles() pattern
I'm not sure when this pattern was added (when file references were added and could
include multiple files?), but ContestObject.getFiles() is used by a bunch of file
reference properties, none of which are then used.

I don't think it's a good pattern to force all file references of a property to
be downloaded at once. (the playback contest does this, but with targeted
comparisons) It is not consistent or done with any other file references and
just complicates the API, so removing this.