URL account tokens for reaction videos
Adds account-specific token parameters to reaction video downloads. Any client using
the url as-is (without basic auth) can download the file as the given user.
This allows (e.g.) web-based clients to pass the url to players like video.js that
support https but not basic auth.

The token is a one-way hash based on the account user and password. It is only
implemented for reaction videos for now to get some use and feedback, but could be
used for other 'file downloads' as well. Since all traffic is via https it could
probably use a simpler algorithm, but this can easily be changed later.

I moved the existing static method used to determine which account is making any
request to HttpHelper since it is not contest-specific, and added the new token
helpers there.

Example of reaction video in /submissions or event-feed after this change:
{"href":"contests/baku-live/submissions/1/reaction.webcam?token=DSjPuOj7OL9HLxEtMe0a5Mrl1vGCz2N2cO1bidDA=","filename":"reaction.webcam.m2ts","mime":"video/m2ts","tags":["webcam"]}],"contest_time":"0:07:47.704","time":"2025-05-26T09:55:53.702-04:00"}