Avoid websocket reconnect on normal disconnect
When the resolver pings the CDS to confirm if it is an admin or regular client, the
default logic automatically reconnects. Depending on conditions this can cause a race
condition on the CDS and a lot of useless logging that the client is already connected.

This avoids client reconnection on a normal/expected disconnection. I've also updated
the disconnection due to invalid login from UNEXPECTED_CONDITION to CANNOT_ACCEPT and
made sure that the disconnection cleanup due to connection failure correctly returns
uses UNEXPECTED_CONDITION (although in practice the client never sees this).

Fixes #1132.