Add timeout computing localhost and avoid when unnecessary
As highlighted in #1129, if you're behind the GFW or simply on a machine with bad
network configuration, pinging google.com in order to determine localhost can hang
for 15s seconds (or potentially longer, depends on host or JDK configuration).

The first thing we can do is put a timeout on the ping. I went with 2s, because if
you can't get to google.com in 2s, you're probably not going to and we should
fallback to trying other options.

Next, we can avoid the call entirely when it's unnecessary. We could try every
tool, but just starting with the resolver running a local feed I found the
following:
 - Trace - this was put here for debugging purposes but we've only used it
   extremely rarely so I think we can just remove. (formatter also applied itself)
 - JSONEncoder - this is only needed on the CDS for file references. We can defer
   the call until it's actually required.
 - NoPresentation - this can be handy to debug clients, but no need to do it on
   the main thread, so I switch to a background call.