Remove support for hidden groups
The 2020-03 spec allowed groups to be hidden. In 2021-11 this property was moved to
teams. The CDS maintained backward compatibility through the PlaybackContest - if you
pointed at a 2020-03 source it would automatically mark any teams in a hidden group
as hidden for any 2021-11 clients.

It's time to remove this support, since the conversion is already complicated and
the current draft spec is changing how this works again.

Also remove unused isJudgementHidden() from ConfiguredContest, and fix javadoc in
IContest (although the draft spec will likely make it wrong again shortly).